refactor: relocate training list utilities
