+++ src/components/training/DragPreviewItem.tsx
