+++ src/components/training/VisibilityPanel.tsx
